Added component files and began work on live camera feed
